mroe cleanup, move 1 usage of RenderPost into the class that uses it
