+++ TASVideos/Pages/Forum/Topics/PollResults.cshtml
@@ -8,7 +8,7 @@
 <hr />
 
 <div>
-	@Html.Raw(Model.Poll.Question)
+	<forum-markup markup=@Model.Poll.Question enable-bb-code=true></forum-markup>
 </div>
 <hr />
 <table class="table table-bordered table-striped">
